+++ NEWS.md
@@ -10,6 +10,9 @@ Fixes:
   * Prevent crash while printing circular plots for the reports on Macs where the svg engines are not working.
   * Prevent crash in residency() if one group has no valid detections at all.
 
+Changes:
+  * Removed deprecated argument `plot.detections.by`.
+
 Enhancements:
   * New `Code.spaces` column in the biometrics allows the user to specificy the code space of the target tags.
   * New arguments in plotArray(): `by.group` and `y.style`. See function documentation for more details.